requirements.txt
# Discord
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0

# LLM
ollama>=0.1.0
aiohttp>=3.9.0

# Text-to-Speech
edge-tts>=6.1.0

# RVC (Voice Conversion)
# Note: RVC requires additional setup - see README
# We'll use so-vits-svc or rvc-python
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
torch>=2.0.0
torchaudio>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
